Extract fetchNearbyProtests to api file Fetch nearby protests on map move end (instead of on every move)
