+++ src/components/ProtestForm/ProtestForm.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
 // import { ReCaptcha, loadReCaptcha } from 'react-recaptcha-v3';
@@ -7,12 +7,8 @@ import { useForm } from 'react-hook-form';
 import { Map, TileLayer, Marker } from 'react-leaflet';
 import Button from '../Button';
 import { validateLatLng } from '../../utils';
-import { createPendingProtest } from '../../api';
-import firebase, { firestore } from '../../firebase';
-import * as geofirestore from 'geofirestore';
+import { createPendingProtest, fetchNearbyProtests } from '../../api';
 import L from 'leaflet';
-
-const GeoFirestore = geofirestore.initializeApp(firestore);
 
 const protestMarker = new L.Icon({
   iconUrl: '/icons/black-flag.svg',
@@ -21,32 +17,20 @@ const protestMarker = new L.Icon({
   iconAnchor: [25, 48],
 });
 
-const geocollection = GeoFirestore.collection('protests');
-const getNearProtests = async (position) => {
-  const query = geocollection.near({
-    center: new firebase.firestore.GeoPoint(position[0], position[1]),
-    radius: 2,
-  });
-  const snapshot = await query.limit(10).get();
-  const protests = snapshot.docs.map((doc) => {
-    const { latitude, longitude } = doc.data().g.geopoint;
-    const protestLatlng = [latitude, longitude];
-    return {
-      id: doc.id,
-      latlng: protestLatlng,
-      ...doc.data(),
-    };
-  });
-  return protests;
-};
-
 function ProtestForm({ initialCoords }) {
-  const { register, handleSubmit } = useForm();
-  const [coordinates, setCoordinates] = useState(() => {
+  const coordinatesUpdater = useCallback(() => {
     let initialState = [31.7749837, 35.219797];
     if (validateLatLng(initialCoords)) initialState = initialCoords;
     return initialState;
-  });
+  }, [initialCoords]);
+
+  const { register, handleSubmit } = useForm();
+  // These two are separate so that onMoveEnd isn't called on every map move
+  // map center
+  const [coordinates, setCoordinates] = useState(coordinatesUpdater);
+  // position of marker
+  const [markerPostion, setMarkerPosition] = useState(coordinatesUpdater);
+
   const [streetName, setStreetName] = useState('');
   // const [recaptchaToken, setRecaptchaToken] = useState('');
   const [submitSuccess, setSubmitSuccess] = useState(false);
@@ -55,13 +39,15 @@ function ProtestForm({ initialCoords }) {
   const [zoomLevel, setZoomLevel] = useState(14);
   // const { recaptcha } = useRef(null);
 
+  // Load nearby protests on mount
   useEffect(() => {
+    const coords = coordinatesUpdater();
     async function nearbyProtests() {
-      const protests = await getNearProtests(coordinates);
+      const protests = await fetchNearbyProtests(coords);
       setNearbyProtests(protests);
     }
     nearbyProtests();
-  }, [streetName]);
+  }, [coordinatesUpdater]);
 
   const onSubmit = async (params) => {
     if (!streetName) {
@@ -127,10 +113,19 @@ function ProtestForm({ initialCoords }) {
             zoom={zoomLevel}
             scrollWheelZoom={'center'}
             onMove={(t) => {
-              setCoordinates([t.target.getCenter().lat, t.target.getCenter().lng]);
+              setMarkerPosition([t.target.getCenter().lat, t.target.getCenter().lng]);
               setZoomLevel(t.target._zoom);
             }}
-            onZoom={(event)=>{
+            onMoveEnd={async (t) => {
+              const newPosition = [t.target.getCenter().lat, t.target.getCenter().lng];
+              setCoordinates(newPosition);
+              setMarkerPosition(newPosition);
+              setZoomLevel(t.target._zoom);
+              // fetch protests on move end
+              const protests = await fetchNearbyProtests(coordinates);
+              setNearbyProtests(protests);
+            }}
+            onZoom={(event) => {
               setZoomLevel(event.target._zoom);
             }}
           >
@@ -138,7 +133,7 @@ function ProtestForm({ initialCoords }) {
               attribution='&amp;copy <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
               url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
             />
-            <Marker position={coordinates}></Marker>
+            <Marker position={markerPostion}></Marker>
             {nearbyProtests.map((protest) => (
               <Marker position={protest.latlng} icon={protestMarker} key={protest.id}></Marker>
             ))}
